Keep chorus depth and delay fractional
depthSamples and delaySamples were stored as unsigned int and rounded on
assignment, but the chorus read position is interpolated: the process
loop computes delay + lfo * depth in floating point and linearly
interpolates between the two neighbouring samples. Rounding the
parameters quantizes the modulation itself, which is what those
fractional sample counts exist to avoid.

Upstream keeps both as plain numbers and only clamps delaySamples to a
minimum of 1. At 48 kHz full-scale depth is 1200 samples, so the error is
small at large depths but reaches several percent at the small ones the
GS chorus presets actually use.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/chorus.h
@@ -43,8 +43,10 @@ typedef struct {
 	float gain;
 	float reverbGain;
 	float delayGain;
-	unsigned int depthSamples;
-	unsigned int delaySamples;
+	/* Fractional sample counts: the read position is interpolated, so
+	 * rounding these to whole samples quantizes the modulation itself. */
+	float depthSamples;
+	float delaySamples;
 	float rateInc;
 	float feedbackGain;
 } SS_Chorus;
